+++ README.md
@@ -9,3 +9,5 @@ Tasks
  * Multi language support
  * Slack bot 
  * wals data features 
+
+To use fastText hindi word vectors with spacy, follow this blog https://medium.com/@expanded_blue_elk_810/rasa-loading-fasttext-vectors-with-spacy-1aad2a2431b
