--- action_utterance_template.py
@@ -1,20 +1,0 @@
- 
-templates = {
-    "action_greet": {
-        "en": {"latin": "Hey! How are you?", "devanagari": "हेय हऊ आर यू"},
-        "hi": {"latin": "Namaste! Kaise ho aap", "devanagari": "नमस्ते। कैसे हो आप"}
-    },
-    
-    "action_did_that_help": {
-        "en": {"latin": "Did that help you?", "devanagari": "डिड थत हेल्प यू"},
-        "hi": {"latin": "Kuch asar hua?", "devanagari": "कुछ असर हुआ?"}
-    },
-    "action_happy": {
-        "en": {"latin": "Great carry on!", "devanagari": "ग्रेट करी ओन!"},
-        "hi": {"latin": "lage raho!", "devanagari": "लगे रहो!"}
-    },
-    "action_goodbye": {
-        "en": {"latin": "Bye", "devanagari": "बाय"},
-        "hi": {"latin": "chalo!", "devanagari": "चलो!"}
-    }
-}
