+++ data/nlu.yml
@@ -29,18 +29,6 @@ nlu:
     - ऐसा नहीं है
     - बिल्कुल नहीं
     - ज़रुरी नहीं
-
-- intent: mood_great
-  examples: |
-    - उत्तम
-    - बहुत अच्छा
-    - महान
-    - गजब का
-    - आश्चर्यजनक
-    - मुझे बहुत अच्छा लग रहा है
-    - मैं महान हूं
-    - मैं अच्छा हूँ
-    - मै ठीक हु 
     
 - intent: thanks
   examples: |
@@ -49,5 +37,9 @@ nlu:
 
 - intent: lang_search
   examples: |
-    - मुझे [हिंदी](language) के बारे जानकारी चाहिए 
+    - मुझे [हिंदी](language) के बारे में जानकारी चाहिए 
     - [उर्दू](language) के बारे में बताओ
+    - मैं [फ्रेंच](language) के बारे में जानना चाहता हूं
+    - आप [स्पेनिश](language) के बारे में क्या जानते हैं
+    - [चीनी](language) भाषा के बारे में बताओx
+    - मुझे [उर्दू](language) के बारे जानकारी चाहिए
