+++ domain.yml
@@ -1,7 +1,6 @@
- 
 version: '2.0'
 session_config:
-  session_expiration_time: 60
+  session_expiration_time: 120
   carry_over_slots_to_new_session: true
 intents:
   - greet
@@ -10,12 +9,10 @@ intents:
   - affirm
   - deny
   - lang_search
+
 entities:
   - language
-# slots:
-#   email:
-#     type: unfeaturized
-#     influence_conversation: false
+
 responses:
   utter_greet:
   - text: |
@@ -24,6 +21,7 @@ responses:
       नमस्ते !
   - text: |
       मैं आपके लिया क्या कर सकता हु ? 
+
   utter_goodbye:
   - text: |
       अलविदा 
@@ -31,6 +29,7 @@ responses:
       आप से बात करके अच्छा लगा 
   - text: |
       फिर मिलते है 
+  
   utter_thanks:
   - text: |
       अच्छा लगा सुनकर मै आपकी मदद कर पाया  
@@ -40,6 +39,7 @@ responses:
   utter_feedback_positive:
   - text: |
       सहायता करके मुझे खुशी हुई :) 
+  
   utter_feedback_negetive:
   - text: |
       क्षमा करें, मैं अगली बार बेहतर करने की कोशिश करूंगा :(
